+++ NEWS.md
@@ -1,8 +1,11 @@
+# desplot 1.11 - unpublished
+
+* Switch to MIT license.
+
 # desplot 1.10 - Mar 2023
 
 * One-row panels no longer have whitespace. Issue #9.
 * Replace `aes_string(x=x.string)` with `aes(x=.data[[x.string]])` etc.
-
 
 # desplot 1.9 - Oct 2021